requirements.txt
google-generativeai
duckduckgo-search
python-dotenv
rich
fastapi
uvicorn
pydantic
